implemented investment service
