+++ 01-starting-project/src/app/app.component.ts
@@ -1,8 +1,7 @@
-import { Component, signal } from '@angular/core';
+import { Component } from '@angular/core';
 
 import { HeaderComponent } from "./header/header.component";
 import { UserInputComponent } from "./user-input/user-input.component";
-import { type InvestmentInput } from './investment-input.model';
 import { InvestmentResultsComponent } from "./investment-results/investment-results.component";
 
 @Component({
@@ -13,36 +12,4 @@ import { InvestmentResultsComponent } from "./investment-results/investment-resu
 })
 
 export class AppComponent {
-  resultsData = signal<{
-    year: number;
-    interest: number;
-    valueEndOfYear: number;
-    annualInvestment: number;
-    totalInterest: number;
-    totalAmountInvested: number;
-  }[] | undefined>(undefined);
-
-  onCalculateInvestmentResults(data: InvestmentInput) {
-    const {initialInvestment, annualInvestment, expectedReturn, duration} = data;
-    const annualData = [];
-    let investmentValue = initialInvestment;
-  
-    for (let i = 0; i < duration; i++) {
-      const year = i + 1;
-      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-      investmentValue += interestEarnedInYear + annualInvestment;
-      const totalInterest =
-        investmentValue - annualInvestment * year - initialInvestment;
-      annualData.push({
-        year: year,
-        interest: interestEarnedInYear,
-        valueEndOfYear: investmentValue,
-        annualInvestment: annualInvestment,
-        totalInterest: totalInterest,
-        totalAmountInvested: initialInvestment + annualInvestment * year,
-      });
-    }
-  
-    this.resultsData.set(annualData);
-  }
 }
