+++ 01-starting-project/src/app/investment-results/investment-results.component.html
@@ -1,4 +1,4 @@
-@if (!results()) {
+@if (!results) {
     <p class="center">Please enter some values and press "Calculate".</p>
 } @else {
     <table>
@@ -12,7 +12,7 @@
             </tr>
         </thead>
         <tbody>
-            @for (result of results(); track result.year) {
+            @for (result of results; track result.year) {
                 <tr>
                     <td>{{ result.year }}</td>
                     <td>{{ result.valueEndOfYear | currency }}</td>
